requirements.txt
pycurl
configparser
GitPython